+++ lithops/executors.py
@@ -1,7 +1,6 @@
 #
-# Copyright 2018 PyWren Team
-# Copyright IBM Corp. 2020
-# Copyright Cloudlab URV 2020
+# (C) Copyright IBM Corp. 2020
+# (C) Copyright Cloudlab URV 2020
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -27,7 +26,8 @@ from lithops.wait import wait_storage, wait_rabbitmq, ALL_COMPLETED
 from lithops.job import create_map_job, create_reduce_job, clean_job
 from lithops.config import default_config, extract_storage_config, \
     default_logging_config, extract_localhost_config, \
-    extract_standalone_config, extract_serverless_config
+    extract_standalone_config, extract_serverless_config, LOCALHOST,\
+    SERVERLESS, STANDALONE
 from lithops.utils import timeout_handler, is_notebook, \
     is_unix_system, is_lithops_worker, create_executor_id
 from lithops.localhost.localhost import LocalhostHandler
@@ -39,31 +39,41 @@ logger = logging.getLogger(__name__)
 
 class FunctionExecutor:
     """
-    Executor asbtract class that contains the common logic
+    Executor abstract class that contains the common logic
     for the Localhost, Serverless and Standalone executors
     """
 
-    def __init__(self, type=None, config=None, backend=None, storage=None,
+    def __init__(self, type=None, mode=None, config=None, backend=None, storage=None,
                  runtime=None, runtime_memory=None, rabbitmq_monitor=None,
                  workers=None, remote_invoker=None, log_level=None):
 
-        if type is None:
+        mode = mode or type
+
+        if mode is None:
             config = default_config(copy.deepcopy(config))
-            type = config['lithops']['executor']
+            mode = config['lithops']['mode']
+
+        if mode not in [LOCALHOST, SERVERLESS, STANDALONE]:
+            raise Exception("Function executor mode must be one of '{}', '{}' "
+                            "or '{}'".format(LOCALHOST, SERVERLESS, STANDALONE))
 
         if log_level:
             default_logging_config(log_level)
 
-        config_ow = {'lithops': {'executor': type}, type: {}}
+        if type is not None:
+            logger.warning("'type' parameter is deprecated and it will be removed"
+                           "in future releases. Use 'mode' parameter instead")
+
+        config_ow = {'lithops': {'mode': mode}, mode: {}}
 
         if runtime is not None:
-            config_ow[type]['runtime'] = runtime
+            config_ow[mode]['runtime'] = runtime
         if backend is not None:
-            config_ow[type]['backend'] = backend
+            config_ow[mode]['backend'] = backend
         if runtime_memory is not None:
-            config_ow[type]['runtime_memory'] = int(runtime_memory)
+            config_ow[mode]['runtime_memory'] = int(runtime_memory)
         if remote_invoker is not None:
-            config_ow[type]['remote_invoker'] = remote_invoker
+            config_ow[mode]['remote_invoker'] = remote_invoker
 
         if storage is not None:
             config_ow['lithops']['storage'] = storage
@@ -88,8 +98,8 @@ class FunctionExecutor:
                 raise Exception("You cannot use rabbitmq_mnonitor since "
                                 "'amqp_url' is not present in configuration")
 
-        self.storage_config = extract_storage_config(self.config)
-        self.internal_storage = InternalStorage(self.storage_config)
+        storage_config = extract_storage_config(self.config)
+        self.internal_storage = InternalStorage(storage_config)
         self.storage = self.internal_storage.storage
 
         self.futures = []
@@ -97,7 +107,7 @@ class FunctionExecutor:
         self.cleaned_jobs = set()
         self.last_call = None
 
-        if type == 'localhost':
+        if mode == LOCALHOST:
             localhost_config = extract_localhost_config(self.config)
             self.compute_handler = LocalhostHandler(localhost_config)
 
@@ -105,16 +115,16 @@ class FunctionExecutor:
                                              self.executor_id,
                                              self.internal_storage,
                                              self.compute_handler)
-        elif type == 'serverless':
+        elif mode == SERVERLESS:
             serverless_config = extract_serverless_config(self.config)
             self.compute_handler = ServerlessHandler(serverless_config,
-                                                     self.storage_config)
+                                                     storage_config)
 
             self.invoker = ServerlessInvoker(self.config,
                                              self.executor_id,
                                              self.internal_storage,
                                              self.compute_handler)
-        elif type == 'standalone':
+        elif mode == STANDALONE:
             standalone_config = extract_standalone_config(self.config)
             self.compute_handler = StandaloneHandler(standalone_config)
 
@@ -122,12 +132,9 @@ class FunctionExecutor:
                                              self.executor_id,
                                              self.internal_storage,
                                              self.compute_handler)
-        else:
-            raise Exception("Function executor type must be one of "
-                            "'localhost', 'serverless' or 'standalone'")
 
-        logger.info('{} Executor created with ID: {}'.format(type.capitalize(),
-                                                             self.executor_id))
+        logger.info('{} Executor created with ID: {}'
+                    .format(mode.capitalize(), self.executor_id))
 
     def __enter__(self):
         return self
@@ -144,12 +151,13 @@ class FunctionExecutor:
 
         :param func: the function to map over the data
         :param data: input data
-        :param extra_data: Additional data to pass to action. Default None.
-        :param extra_env: Additional environment variables for action environment. Default None.
-        :param runtime_memory: Memory to use to run the function. Default None (loaded from config).
-        :param timeout: Time that the functions have to complete their execution before raising a timeout.
-        :param include_modules: Explicitly pickle these dependencies.
-        :param exclude_modules: Explicitly keep these modules from pickled dependencies.
+        :param extra_env: Additional env variables for action environment
+        :param runtime_memory: Memory to use to run the function
+        :param timeout: Time that the functions have to complete their
+                        execution before raising a timeout
+        :param include_modules: Explicitly pickle these dependencies
+        :param exclude_modules: Explicitly keep these modules from pickled
+                                dependencies
 
         :return: future object.
         """
@@ -174,24 +182,29 @@ class FunctionExecutor:
 
         return futures[0]
 
-    def map(self, map_function, map_iterdata, extra_args=None, extra_env=None, runtime_memory=None,
-            chunk_size=None, chunk_n=None, timeout=None, invoke_pool_threads=500,
-            include_modules=[], exclude_modules=[]):
+    def map(self, map_function, map_iterdata, extra_args=None, extra_env=None,
+            runtime_memory=None, chunk_size=None, chunk_n=None, timeout=None,
+            invoke_pool_threads=500, include_modules=[], exclude_modules=[]):
         """
+        For running multiple function executions asynchronously
+
         :param map_function: the function to map over the data
         :param map_iterdata: An iterable of input data
-        :param extra_args: Additional arguments to pass to the function activation. Default None.
-        :param extra_env: Additional environment variables for action environment. Default None.
-        :param runtime_memory: Memory to use to run the function. Default None (loaded from config).
-        :param chunk_size: the size of the data chunks to split each object. 'None' for processing
-                           the whole file in one function activation.
-        :param chunk_n: Number of chunks to split each object. 'None' for processing the whole
-                        file in one function activation.
-        :param remote_invocation: Enable or disable remote_invocation mechanism. Default 'False'
-        :param timeout: Time that the functions have to complete their execution before raising a timeout.
-        :param invoke_pool_threads: Number of threads to use to invoke.
-        :param include_modules: Explicitly pickle these dependencies.
-        :param exclude_modules: Explicitly keep these modules from pickled dependencies.
+        :param extra_args: Additional args to pass to the function activations
+        :param extra_env: Additional env variables for action environment
+        :param runtime_memory: Memory to use to run the function
+        :param chunk_size: the size of the data chunks to split each object.
+                           'None' for processing the whole file in one function
+                           activation.
+        :param chunk_n: Number of chunks to split each object. 'None' for
+                        processing the whole file in one function activation
+        :param remote_invocation: Enable or disable remote_invocation mechanism
+        :param timeout: Time that the functions have to complete their execution
+                        before raising a timeout
+        :param invoke_pool_threads: Number of threads to use to invoke
+        :param include_modules: Explicitly pickle these dependencies
+        :param exclude_modules: Explicitly keep these modules from pickled
+                                dependencies
 
         :return: A list with size `len(iterdata)` of futures.
         """
@@ -220,8 +233,9 @@ class FunctionExecutor:
 
         return futures
 
-    def map_reduce(self, map_function, map_iterdata, reduce_function, extra_args=None, extra_env=None,
-                   map_runtime_memory=None, reduce_runtime_memory=None, chunk_size=None, chunk_n=None,
+    def map_reduce(self, map_function, map_iterdata, reduce_function,
+                   extra_args=None, extra_env=None, map_runtime_memory=None,
+                   reduce_runtime_memory=None, chunk_size=None, chunk_n=None,
                    timeout=None, invoke_pool_threads=500, reducer_one_per_object=False,
                    reducer_wait_local=False, include_modules=[], exclude_modules=[]):
         """
@@ -298,8 +312,9 @@ class FunctionExecutor:
 
         return map_futures + reduce_futures
 
-    def wait(self, fs=None, throw_except=True, return_when=ALL_COMPLETED, download_results=False,
-             timeout=None, THREADPOOL_SIZE=128, WAIT_DUR_SEC=1):
+    def wait(self, fs=None, throw_except=True, return_when=ALL_COMPLETED,
+             download_results=False, timeout=None, THREADPOOL_SIZE=128,
+             WAIT_DUR_SEC=1):
         """
         Wait for the Future instances (possibly created by different Executor instances)
         given by fs to complete. Returns a named 2-tuple of sets. The first set, named done,
@@ -386,7 +401,7 @@ class FunctionExecutor:
                 print()
             logger.info(msg)
             if not self.log_active:
-                print(msg) 
+                print(msg)
             error = True
             raise e
 
@@ -416,7 +431,8 @@ class FunctionExecutor:
 
         return fs_done, fs_notdone
 
-    def get_result(self, fs=None, throw_except=True, timeout=None, THREADPOOL_SIZE=128, WAIT_DUR_SEC=1):
+    def get_result(self, fs=None, throw_except=True, timeout=None,
+                   THREADPOOL_SIZE=128, WAIT_DUR_SEC=1):
         """
         For getting the results from all function activations
 
@@ -429,22 +445,25 @@ class FunctionExecutor:
 
         :return: The result of the future/s
         """
-        fs_done, unused_fs_notdone = self.wait(fs=fs, throw_except=throw_except,
-                                               timeout=timeout, download_results=True,
-                                               THREADPOOL_SIZE=THREADPOOL_SIZE,
-                                               WAIT_DUR_SEC=WAIT_DUR_SEC)
+        fs_done, _ = self.wait(fs=fs, throw_except=throw_except,
+                               timeout=timeout, download_results=True,
+                               THREADPOOL_SIZE=THREADPOOL_SIZE,
+                               WAIT_DUR_SEC=WAIT_DUR_SEC)
         result = []
         fs_done = [f for f in fs_done if not f.futures and f._produce_output]
         for f in fs_done:
             if fs:
                 # Process futures provided by the user
-                result.append(f.result(throw_except=throw_except, internal_storage=self.internal_storage))
+                result.append(f.result(throw_except=throw_except,
+                                       internal_storage=self.internal_storage))
             elif not fs and not f._read:
                 # Process internally stored futures
-                result.append(f.result(throw_except=throw_except, internal_storage=self.internal_storage))
+                result.append(f.result(throw_except=throw_except,
+                                       internal_storage=self.internal_storage))
                 f._read = True
 
-        logger.debug("ExecutorID {} Finished getting results".format(self.executor_id))
+        logger.debug("ExecutorID {} Finished getting results"
+                     .format(self.executor_id))
 
         if len(result) == 1 and self.last_call != 'map':
             return result[0]
@@ -467,7 +486,8 @@ class FunctionExecutor:
         ftrs_to_plot = [f for f in ftrs if (f.ready or f.done) and not f.error]
 
         if not ftrs_to_plot:
-            logger.debug('ExecutorID {} - No futures ready to plot'.format(self.executor_id))
+            logger.debug('ExecutorID {} - No futures ready to plot'
+                         .format(self.executor_id))
             return
 
         logging.getLogger('matplotlib').setLevel(logging.WARNING)
@@ -499,7 +519,8 @@ class FunctionExecutor:
             futures = [futures]
 
         if not futures:
-            logger.debug('ExecutorID {} - No jobs to clean'.format(self.executor_id))
+            logger.debug('ExecutorID {} - No jobs to clean'
+                         .format(self.executor_id))
             return
 
         if fs or force:
@@ -517,11 +538,15 @@ class FunctionExecutor:
             if not self.log_active:
                 print(msg)
             storage_config = self.internal_storage.get_storage_config()
-            clean_job(jobs_to_clean, storage_config, clean_cloudobjects=cloudobjects)
+            clean_job(jobs_to_clean, storage_config,
+                      clean_cloudobjects=cloudobjects)
             self.cleaned_jobs.update(jobs_to_clean)
 
     def dismantle(self):
         self.compute_handler.dismantle()
+
+    def init(self):
+        self.compute_handler.init()
 
     def __exit__(self, exc_type, exc_value, traceback):
         self.invoker.stop()
@@ -545,7 +570,7 @@ class LocalhostExecutor(FunctionExecutor):
 
         :return `LocalhostExecutor` object.
         """
-        super().__init__(type='localhost', config=config,
+        super().__init__(mode=LOCALHOST, config=config,
                          runtime=runtime, storage=storage,
                          log_level=log_level, workers=workers,
                          rabbitmq_monitor=rabbitmq_monitor)
@@ -570,7 +595,7 @@ class ServerlessExecutor(FunctionExecutor):
 
         :return `ServerlessExecutor` object.
         """
-        super().__init__(type='serverless', config=config, runtime=runtime,
+        super().__init__(mode=SERVERLESS, config=config, runtime=runtime,
                          runtime_memory=runtime_memory, backend=backend,
                          storage=storage, workers=workers,
                          rabbitmq_monitor=rabbitmq_monitor, log_level=log_level,
@@ -594,6 +619,6 @@ class StandaloneExecutor(FunctionExecutor):
 
         :return `StandaloneExecutor` object.
         """
-        super().__init__(type='standalone', config=config, runtime=runtime,
+        super().__init__(mode=STANDALONE, config=config, runtime=runtime,
                          backend=backend, storage=storage, workers=workers,
                          rabbitmq_monitor=rabbitmq_monitor, log_level=log_level)